+++ packages/react/src/errorBoundary/types.ts
@@ -1,5 +1,7 @@
 import type { ReactElement, ReactNode } from 'react';
 
+import type { PushErrorOptions } from '@grafana/faro-web-sdk';
+
 export type ReactNodeRender = () => ReactNode;
 
 export type ReactProps = Record<string, any>;
@@ -14,6 +16,7 @@ export interface FaroErrorBoundaryProps {
   onMount?: VoidFunction;
   onReset?: (error: Error | null) => void;
   onUnmount?: (error: Error | null) => void;
+  pushErrorOptions?: PushErrorOptions;
 }
 
 export interface FaroErrorBoundaryState {